refactored render-world for multiple segments
Commented out the majority of the code, as very little of it is able to
work with lists. I added render-segment, which does what the earlier
render-world did. Added render-worm, which is a recursive version of
render-segment. Changed render-world to break up the game state into
worm and food, then control the other rendering functions. Refer to
TO-DO at the bottom of the file for further refactoring work.